requirements.txt
faiss-gpu
numpy
Pillow
scikit-learn
torch
torchaudio
torchvision
wandb
